Move Lvup into Kh2.Battle
